# Conflicts:
#	pkgs/native_toolchain_c/CHANGELOG.md
#	pkgs/native_toolchain_c/pubspec.yaml

+++ pkgs/jnigen/lib/src/tools/android_sdk_tools.dart
@@ -4,6 +4,7 @@
 
 import 'dart:io';
 
+import 'package:jni_util/jni_util.dart';
 import 'package:path/path.dart';
 import 'package:yaml/yaml.dart';
 
@@ -482,7 +483,9 @@ tasks.register<DefaultTask>("$_gradleGetSourcesTaskName") {
     ProcessResult procRes;
     try {
       procRes = Process.runSync(gradleCommand, ['-q', taskPath],
-          workingDirectory: android, runInShell: true);
+          workingDirectory: android,
+          runInShell: true,
+          environment: javaEnvironment);
     } finally {
       log.info('Restoring build scripts');
       origBuild.writeAsStringSync(
@@ -490,7 +493,9 @@ tasks.register<DefaultTask>("$_gradleGetSourcesTaskName") {
             .replaceAll(_gradleGetClasspathStub, '')
             .replaceAll(_gradleGetSourcesStub, ''),
       );
-      File(buildGradleOld).deleteSync();
+      if (File(buildGradleOld).existsSync()) {
+        File(buildGradleOld).deleteSync();
+      }
     }
     if (procRes.exitCode != 0) {
       final inAndroidProject = _inAndroidProject(androidProject);
@@ -515,8 +520,8 @@ Execution failed for task ':gradle:compileGroovy'.
 > BUG! exception in phase 'semantic analysis' ...  Unsupported class file major version
 
 Then the JDK versions used by JNIgen and Flutter are not compatible. Try
-changing the default JDK version e.g. with `export JAVA_VERSION=11` on macOS and
-`sudo update-alternatives --config java` on Ubuntu.
+changing the default JDK version e.g. with `export JAVA_VERSION=$minJavaVersion`
+on macOS and `sudo update-alternatives --config java` on Ubuntu.
 
 GRADLE OUTPUT:
 --------------------------------------------------------------------------------
@@ -561,6 +566,7 @@ ${procRes.stderr}
         ['build', 'apk', '--config-only'],
         workingDirectory: androidProject,
         runInShell: true,
+        environment: javaEnvironment,
       );
       if (result.exitCode == 0) {
         success = true;
